Refactoring

+++ app/handlers/events.py
@@ -45,6 +45,7 @@ def my_id_name(message: Message):
     return message
 
 def my_id_name_from_callback(callback: CallbackQuery):
+    """Generates a message containing the user's name and ID from a callback query."""
     user = callback.from_user
     user_name = user.full_name
     user_id = user.id
@@ -60,17 +61,15 @@ def is_url(text):
 
 def generate_image(prompt):
     """Generates an image based on a given prompt using the TextToImageAPI class."""
-    imgGen = TextToImageAPI()
-    result = imgGen.generate_image(prompt)
-    return result
+    return TextToImageAPI().generate_image(prompt)
 
 async def cmd_gen_image(message: Message):
     """Generates and sends an image based on a prompt."""
     reply_message = await message.reply(f"{MESSAGES['AWAITING']}")
-
+    
     prompt = promt_generator()
     api_requests = generate_image(prompt)
-
+    
     if is_url(api_requests):
         caption = f"{MESSAGES['PROMT']}{hbold(prompt)}"
         await message.answer_photo(photo=api_requests, caption=caption)
@@ -82,17 +81,19 @@ async def cmd_gen_image(message: Message):
 async def callback_gen_image(callback: CallbackQuery):
     """Function to handle the callback query with data 'GEN' from the router."""
     reply_message = await callback.message.answer(f"{MESSAGES['AWAITING']}")
-
+    
     prompt = promt_generator()
     api_requests = generate_image(prompt)
-
+    
     if is_url(api_requests):
         caption = f"{MESSAGES['PROMT']}{hbold(prompt)}"
-        await callback.message.answer_photo(photo=api_requests, caption=caption, reply_markup=after_gen_keyboard)
+        photo_message = await callback.message.answer_photo(photo=api_requests, caption=caption, reply_markup=after_gen_keyboard)
     else:
         await callback.answer(api_requests, reply_markup=after_gen_keyboard)
-
-    await bot.delete_message(callback.from_user.id, reply_message.message_id)
+        photo_message = None
+    
+    if photo_message:
+        await bot.delete_message(callback.from_user.id, reply_message.message_id)
 
 async def callback_gen_image_repeat(callback: CallbackQuery):
     """Function to handle the callback query with data 'GEN' from the router."""
@@ -102,22 +103,27 @@ async def callback_gen_image_repeat(callback: CallbackQuery):
     await bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id, reply_markup=None)
     
     reply_message = await callback.message.answer(f"{MESSAGES['AWAITING']}")
-
     prompt = promt_generator()
     api_requests = generate_image(prompt)
-
+    
     if is_url(api_requests):
         caption = f"{MESSAGES['PROMT']}{hbold(prompt)}"
-        await callback.message.answer_photo(photo=api_requests, caption=caption, reply_markup=after_gen_keyboard)
+        await callback.message.answer_photo(photo=api_requests, caption=caption,
+                                            reply_markup=after_gen_keyboard)
     else:
         await callback.answer(api_requests, reply_markup=after_gen_keyboard)
-
+    
     await bot.delete_message(callback.from_user.id, reply_message.message_id)
 
 async def callback_back_menu(callback: CallbackQuery):
-    chat_id = callback.message.chat.id
-    message_id = callback.message.message_id
-   
-    await bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id, reply_markup=None)
-   
-    await callback.message.answer(f"{hbold(MESSAGES['MAIN_MENU'])}", reply_markup=start_keyboard)
+    """Asynchronous function that handles the callback from the back menu button."""
+    await bot.edit_message_reply_markup(
+        chat_id=callback.message.chat.id,
+        message_id=callback.message.message_id,
+        reply_markup=None
+    )
+
+    await callback.message.answer(
+        f"{hbold(MESSAGES['MAIN_MENU'])}",
+        reply_markup=start_keyboard
+    )
